OpenGL renderer:  - Implemented ProgramBuffer.  - Emulated more rsx commands.
rpcs3:
 - Fixed crash on stop.

Interpreter & DisAsm:
 - Added Pause button.
 - Fixed freeze on Run.

tetris:
 - Updated to last version.

+++ trunk/rpcs3/Emu/Cell/PPCThread.h
@@ -92,7 +92,7 @@ public:
 	void SetJoining(bool joining) { m_joining = joining; }
 
 	u32 GetError() const { return m_error; }
-	u16 GetId() const { return m_id; }
+	u32 GetId() const { return m_id; }
 
 	void Reset();
 	void Close();
